Permit open upper bound Guava versions in MANIFEST.MF for OSGi
This makes it possible to load this bundle into an OSGi Framework
which has more recent versions of Guava, which of course are
backward compatible (for the use xtend-lib makes...), instead
of being force to make different Guava versions available.

The lower bound minimal version stays at 14.0.0, as is, unchanged.

Signed-off-by: Michael Vorburger <mike@vorburger.ch>
Signed-off-by: Michael Vorburger <vorburger@redhat.com>